+++ buildSrc/src/main/kotlin/my/CustomPlugin.kt
@@ -1,24 +1,11 @@
 package my
 
-import org.gradle.api.Plugin
+import org.gradle.api.AllOpenPlugin
 import org.gradle.api.Project
-import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.kotlin.dsl.*
 
-open class CustomPluginExtension(project: Project) {
-    var helloWorld by project.objects.property<String>()
-}
 
-/** In a mixed Kotlin-Gradle project,
-instead of writing an extension function of Project,
-we write it as a member function of ProjectKt
-which we make available via a plugin as  `project.kt`
-**/
-open class ProjectKt(val project: Project) {
-    fun doSomethingCool() = println("Hello world from project: ${project.name}")
-}
-
-open class CustomPlugin : Plugin<Project> {
+class CustomPlugin : AllOpenPlugin<Project> {
 
     override fun apply(project: Project) {
         var extension = project.extensions.create(
